pyproject.toml
[project]
name = "drtail-prompt"
version = "0.1.0"
description = "Prompt for DrTail"
readme = "README.md"
authors = [
    { name = "Humphrey Ahn", email = "ahnsv@bc.edu" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
